+++ apps/mobile-wallet/src/screens/LandingScreen.tsx
@@ -16,9 +16,10 @@ You should have received a copy of the GNU Lesser General Public License
 along with the library. If not, see <http://www.gnu.org/licenses/>.
 */
 
+import { useFocusEffect } from '@react-navigation/native'
 import { StackScreenProps } from '@react-navigation/stack'
 import { Canvas, RadialGradient, Rect, vec } from '@shopify/react-native-skia'
-import { useEffect, useState } from 'react'
+import { useCallback, useEffect, useState } from 'react'
 import { Trans, useTranslation } from 'react-i18next'
 import { AppState, Dimensions, Image, LayoutChangeEvent, Platform, StatusBar } from 'react-native'
 import Animated, {
@@ -37,15 +38,15 @@ import styled, { ThemeProvider, useTheme } from 'styled-components/native'
 import AppText from '~/components/AppText'
 import Button from '~/components/buttons/Button'
 import Screen, { ScreenProps } from '~/components/layout/Screen'
-import { useAppDispatch } from '~/hooks/redux'
+import { useAppDispatch, useAppSelector } from '~/hooks/redux'
 import altLogoSrc from '~/images/logos/alephiumHackLogo.png'
 import AlephiumLogo from '~/images/logos/AlephiumLogo'
 import RootStackParamList from '~/navigation/rootStackRoutes'
-import { getWalletMetadata } from '~/persistent-storage/wallet'
+import { storedWalletExists } from '~/persistent-storage/wallet'
 import { methodSelected, WalletGenerationMethod } from '~/store/walletGenerationSlice'
 import { BORDER_RADIUS_BIG, BORDER_RADIUS_HUGE } from '~/style/globalStyle'
 import { themes } from '~/style/themes'
-import { showExceptionToast } from '~/utils/layout'
+import { resetNavigation } from '~/utils/navigation'
 
 interface LandingScreenProps extends StackScreenProps<RootStackParamList, 'LandingScreen'>, ScreenProps {}
 
@@ -57,10 +58,34 @@ const LandingScreen = ({ navigation, ...props }: LandingScreenProps) => {
   const insets = useSafeAreaInsets()
   const theme = useTheme()
   const { t } = useTranslation()
+  const isWalletUnlocked = useAppSelector((s) => s.wallet.isUnlocked)
 
   const { width, height } = Dimensions.get('window')
   const [dimensions, setDimensions] = useState({ width, height })
-  const [showNewWalletButtons, setShowNewWalletButtons] = useState(false)
+  const [isScreenContentVisible, setIsScreenContentVisible] = useState(false)
+
+  useFocusEffect(
+    useCallback(() => {
+      storedWalletExists()
+        .then((walletExists) => {
+          if (walletExists) {
+            // Normally, when the app is unlocked, this screen is not in focus. However, under certain conditions we end
+            // up with an unlocked wallet and no screen in focus at all. This happens when:
+            // 1. the auto-lock is set to anything but "Fast"
+            // 2. the user manually kills the app before the auto-lock timer completes
+            // 3. the WalletConnect feature is activated
+            // Since there is no screen in focus and since the default screen set in the RootStackNavigation is this
+            // screen, we need to navigate back to the dashboard.
+            if (isWalletUnlocked) resetNavigation(navigation)
+          } else {
+            // Only display this screen's contents when we have no stored wallet. If there is a wallet, the
+            // AppUnlockModal will be displayed
+            setIsScreenContentVisible(true)
+          }
+        })
+        .catch((error) => console.error('Could not determine if stored wallet exists', error))
+    }, [isWalletUnlocked, navigation])
+  )
 
   useEffect(() => {
     const unsubscribeBlurListener = navigation.addListener('blur', () => {
@@ -88,16 +113,10 @@ const LandingScreen = ({ navigation, ...props }: LandingScreenProps) => {
     setDimensions({ width, height })
   }
 
-  useEffect(() => {
-    getWalletMetadata()
-      .then((metadata) => setShowNewWalletButtons(!metadata))
-      .catch((e) => showExceptionToast(e, t('Wallet metadata not found')))
-  }, [t])
-
   return (
     <ThemeProvider theme={themes.dark}>
       <Screen contrastedBg {...props} onLayout={handleScreenLayoutChange}>
-        {showNewWalletButtons && (
+        {isScreenContentVisible && (
           <>
             <CoolAlephiumCanvas {...dimensions} />